update on MS reports added filter alla working correcthy

+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/ms/MsComplaintFeedbackViewEntity.kt
@@ -5,7 +5,6 @@ import org.kebs.app.kotlin.apollo.store.model.ms.MsReportSubmittedCpViewEntity
 import org.kebs.app.kotlin.apollo.store.model.ms.MsAcknowledgementTimelineViewEntity
 import org.kebs.app.kotlin.apollo.store.model.ms.MsSampleSubmittedCpViewEntity
 import java.io.Serializable
-import java.math.BigInteger
 import java.sql.Date
 import java.util.*
 import javax.persistence.*
@@ -39,7 +38,7 @@ class MsComplaintFeedbackViewEntity : Serializable {
 
     @Basic
     @Column(name = "ASSIGNED_IO")
-    var assignedIo: BigInteger? = null
+    var assignedIo: Long? = null
 
     @Basic
     @Column(name = "ACKNOWLEDGEMENT_TYPE")
@@ -47,23 +46,23 @@ class MsComplaintFeedbackViewEntity : Serializable {
 
     @Basic
     @Column(name = "REGION")
-    var region: BigInteger? = null
+    var region: Long? = null
 
     @Basic
     @Column(name = "COUNTY")
-    var county: BigInteger? = null
+    var county: Long? = null
 
     @Basic
     @Column(name = "TOWN")
-    var town: BigInteger? = null
+    var town: Long? = null
 
     @Basic
     @Column(name = "COMPLAINT_DEPARTMENT")
-    var complaintDepartment: BigInteger? = null
+    var complaintDepartment: Long? = null
 
     @Basic
     @Column(name = "DIVISION")
-    var division: BigInteger? = null
+    var division: Long? = null
 
     @Basic
     @Column(name = "TIME_TAKEN_FOR_ACKNOWLEDGEMENT")
