+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/ms/MsComplaintsInvestigationsViewEntity.kt
@@ -3,7 +3,6 @@ package org.kebs.app.kotlin.apollo.store.model.ms
 import org.kebs.app.kotlin.apollo.store.model.ms.MsComplaintsInvestigationsViewEntity
 import org.kebs.app.kotlin.apollo.store.model.ms.MsPerformanceOfSelectedProductViewEntity
 import java.io.Serializable
-import java.math.BigInteger
 import java.sql.Date
 import java.util.*
 import javax.persistence.*
@@ -37,7 +36,7 @@ class MsComplaintsInvestigationsViewEntity : Serializable {
 
     @Basic
     @Column(name = "ASSIGNED_IO")
-    var assignedIo: BigInteger? = null
+    var assignedIo: Long? = null
 
     @Basic
     @Column(name = "STATUS")
@@ -45,23 +44,23 @@ class MsComplaintsInvestigationsViewEntity : Serializable {
 
     @Basic
     @Column(name = "REGION")
-    var region: BigInteger? = null
+    var region: Long? = null
 
     @Basic
     @Column(name = "COUNTY")
-    var county: BigInteger? = null
+    var county: Long? = null
 
     @Basic
     @Column(name = "TOWN")
-    var town: BigInteger? = null
+    var town: Long? = null
 
     @Basic
     @Column(name = "COMPLAINT_DEPARTMENT")
-    var complaintDepartment: BigInteger? = null
+    var complaintDepartment: Long? = null
 
     @Basic
     @Column(name = "DIVISION")
-    var division: BigInteger? = null
+    var division: Long? = null
 
     @Basic
     @Column(name = "TIME_TAKEN_FOR_ACKNOWLEDGEMENT")
